refactor: standardize types on bool/uint8_t/size_t/stdint, with char for strings
- bool for booleans, uint8_t for byte buffers, size_t/ssize_t for lengths
- in_port_t for ports, socklen_t for address lengths, sa_family_t for families
- char for string buffers (username/address/hostname) to stay compatible with string functions

+++ src/protocol/handshake.h
@@ -10,6 +10,7 @@
 #define _MRS_PROTOCOLS_HANDSHAKE_H_INCLUDED_
 
 #include <netinet/in.h>
+#include <stdint.h>
 
 #include "../define/global.h"
 
@@ -26,7 +27,7 @@ typedef struct {
 	varint_t id_part1, id_part2, nextstate, version;
 	void *address, *signature_data, *username;
 	size_t signature_data_length;
-	unsigned short version_fml;
+	uint8_t version_fml;
 	in_port_t port;
 } p_handshake;
